Reorganized variables in several classes to reduce memory usage by reducing padding.
These classes had a somewhat significant size reduction: (64-bit Linux)
- RomDataPrivate: 100 -> 96 (-4)
- GcnFstPrivate: 104 -> 88 (-16)
- ELFPrivate: 296 -> 280 (-16)

Converted some static classes to namespaces:
- [win32] RpImageWin32
- [libromdata] CtrKeyScrambler

[win32] MessageWidgetPrivate:
- Un-inline the constructor.
- Don't save the USER32 pointer; it's only needed when changing the icon.

RomFields: Rearranged some fields in Field::desc::bitfield
and Field::desc::list_data to reduce padding.
- This causes ABI breakage, but libromdata.so hasn't been released
  yet, so it's okay.

RomMetaData: Expand Property to `int` for future expansion in case
more properties are added later.
- Again, ABI breakage, but better to do it now before the ABI is
  frozen for this SOVERSION than to do it later.

WiiSave: Move key_idx[] and key_status[] to ENABLE_DECRYPTION.
- NOTE: Not really used much...

NintendoDSPrivate::getDefaultLC(): Removed 'inline'.

NCCHReaderPrivate: Changed headers_loaded from uint32_t to uint8_t.

+++ src/librpbase/RomMetaData.hpp
@@ -28,7 +28,7 @@ namespace LibRpBase {
 
 // Properties.
 // This matches KFileMetaData::Property.
-enum class Property : int8_t {
+enum class Property : int {
 	Invalid = -1,
 	FirstProperty = 0,
 	Empty = 0,
@@ -200,7 +200,7 @@ class RomMetaData
 
 		/**
 		 * Get a metadata property.
-		 * @param idx Field index.
+		 * @param idx Metadata index
 		 * @return Metadata property, or nullptr if the index is invalid.
 		 */
 		RP_LIBROMDATA_PUBLIC
@@ -218,7 +218,7 @@ class RomMetaData
 
 		/**
 		 * Reserve space for metadata.
-		 * @param n Desired capacity.
+		 * @param n Desired capacity
 		 */
 		void reserve(int n);
 
@@ -228,7 +228,7 @@ class RomMetaData
 		 * If metadata properties with the same names already exist,
 		 * they will be overwritten.
 		 *
-		 * @param other Source RomMetaData object.
+		 * @param other Source RomMetaData object
 		 * @return Metadata index of the last metadata added.
 		 */
 		int addMetaData_metaData(const RomMetaData *other);
@@ -239,16 +239,16 @@ class RomMetaData
 		 * If a metadata property with the same name already exists,
 		 * it will be overwritten.
 		 *
-		 * @param name Metadata name.
-		 * @param val Integer value.
+		 * @param name Property name
+		 * @param val Integer value
 		 * @return Metadata index, or -1 on error.
 		 */
 		int addMetaData_integer(Property name, int value);
 
 		/**
 		 * Add an unsigned integer metadata property.
-		 * @param name Metadata name.
-		 * @param val Unsigned integer value.
+		 * @param name Property name
+		 * @param val Unsigned integer value
 		 * @return Metadata index, or -1 on error.
 		 */
 		int addMetaData_uint(Property name, unsigned int value);
@@ -259,9 +259,9 @@ class RomMetaData
 		 * If a metadata property with the same name already exists,
 		 * it will be overwritten.
 		 *
-		 * @param name Metadata name.
-		 * @param str String value.
-		 * @param flags Formatting flags.
+		 * @param name Property name
+		 * @param str String value
+		 * @param flags Formatting flags
 		 * @return Metadata index, or -1 on error.
 		 */
 		int addMetaData_string(Property name, const char *str, unsigned int flags = 0);
@@ -272,9 +272,9 @@ class RomMetaData
 		 * If a metadata property with the same name already exists,
 		 * it will be overwritten.
 		 *
-		 * @param name Metadata name.
-		 * @param str String value.
-		 * @param flags Formatting flags.
+		 * @param name Property name
+		 * @param str String value
+		 * @param flags Formatting flags
 		 * @return Metadata index, or -1 on error.
 		 */
 		int addMetaData_string(Property name, const std::string &str, unsigned int flags = 0);
@@ -285,8 +285,8 @@ class RomMetaData
 		 * If a metadata property with the same name already exists,
 		 * it will be overwritten.
 		 *
-		 * @param name Metadata name.
-		 * @param timestamp UNIX timestamp.
+		 * @param name Metadata name
+		 * @param timestamp UNIX timestamp
 		 * @return Metadata index, or -1 on error.
 		 */
 		int addMetaData_timestamp(Property name, time_t timestamp);
